+++ packages/ui/src/widgets/Leaderboard/TopWidget.tsx
@@ -10,20 +10,24 @@ type TopWidgetProps = {
   onPlayAgain: () => void;
   disabled?: boolean;
   tournamentTitle: string;
+  tournamentSubtitle: string;
+  hasActiveTournament?: boolean;
   onTweet?: (score: number) => Promise<{ tweetBody: string }>;
   score?: number;
+  rank?: number;
 };
 
-const WidgetWrapper = styled(ModalWrapper)({
+const WidgetWrapper = styled(ModalWrapper)(({ tournament }: { tournament?: boolean }) => ({
+  padding: '36px 24px 32px 24px',
   width: 490,
   minHeight: 265,
   '@media (max-width: 600px)': {
     width: 'auto',
-    maxHeight: 325,
+    maxHeight: tournament ? '100%' : 325,
   },
-});
+}));
 
-const DaysLeft = styled.div({
+const DaysLeft = styled.div(({ tournament }: { tournament?: boolean }) => ({
   padding: '7px 15px 7px 32px',
   background: 'rgba(133, 70, 183, 1)',
   fontFamily: 'Yapari-SemiBold',
@@ -36,10 +40,20 @@ const DaysLeft = styled.div({
   top: '-80px',
   minWidth: 168,
   textTransform: 'uppercase',
+  ...(tournament
+    ? {
+        padding: '7px 15px',
+        position: 'relative',
+        top: '0',
+        left: '0',
+        width: '151px',
+        margin: '4px auto 14px auto',
+      }
+    : {}),
   '@media (max-width: 600px)': {
-    left: '-30px',
+    left: tournament ? '0px' : '-30px',
   },
-});
+}));
 
 const Text = styled.div({
   marginTop: 81,
@@ -50,19 +64,22 @@ const Text = styled.div({
   },
 });
 
-const UniqueNFT = styled(Typography)({
+const UniqueNFT = styled(Typography)(({ tournament }: { tournament?: boolean }) => ({
   fontFamily: 'Yapari-SemiBold',
   fontWeight: 600,
   fontSize: 24,
+  ...(tournament && {
+    textTransform: 'uppercase',
+  }),
   '@media (max-width: 600px)': {
-    width: 186,
-    height: 72,
+    width: 287,
     fontSize: 19,
+    margin: '0 auto',
   },
-});
+}));
 
-const PlayAgain = styled(Button)({
-  marginTop: '37px!important',
+const PlayAgain = styled(Button)(({ tournament }: { tournament?: boolean }) => ({
+  marginTop: tournament ? '20px!important' : '37px!important',
   maxWidth: 146,
   height: 36,
   minHeight: 36,
@@ -71,16 +88,22 @@ const PlayAgain = styled(Button)({
   borderRadius: 4,
   padding: 0,
   background: 'rgba(243, 39, 88, 1)',
-});
+  ...(tournament && {
+    whiteSpace: 'nowrap',
+    '@media (max-width: 600px)': {
+      marginTop: '14px!important',
+    },
+  }),
+}));
 
 const PlayAgainText = styled(Typography)({
   marginLeft: 6,
   fontSize: 14,
 });
 
-const TweetButton = styled(Button)({
+const TweetButton = styled(Button)(({ tournament }: { tournament?: boolean }) => ({
   background: 'transparent',
-  marginTop: '37px!important',
+  marginTop: tournament ? '20px!important' : '37px!important',
   border: '1px solid #FFFFFF',
   height: 36,
   minHeight: 36,
@@ -89,10 +112,18 @@ const TweetButton = styled(Button)({
   borderRadius: 4,
   '& > div': {
     padding: '0 6px',
+    ...(tournament && {
+      whiteSpace: 'nowrap',
+    }),
   },
-});
+  ...(tournament && {
+    '@media (max-width: 600px)': {
+      marginTop: '14px!important',
+    },
+  }),
+}));
 
-const GamePortalButton = styled(Typography)({
+const GamePortalButton = styled(Typography)(({ tournament }: { tournament?: boolean }) => ({
   cursor: 'pointer',
   width: 'fit-content',
   marginTop: '11px',
@@ -101,7 +132,10 @@ const GamePortalButton = styled(Typography)({
   lineHeight: '15px',
   fontWeight: 400,
   textDecoration: 'underline',
-});
+  ...(tournament && {
+    margin: '12px auto 0 auto',
+  }),
+}));
 
 const NFTImage = styled.img({
   position: 'absolute',
@@ -117,20 +151,63 @@ const NFTImage = styled.img({
   },
 });
 
-const Row = styled.div(({ columns, columnGap }: { columns: string; columnGap: number }) => ({
+const RewardsRow = styled.div({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(3, 1fr)',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  marginTop: '14px',
+  marginBottom: '20px',
+  '& > div': {
+    justifySelf: 'center',
+  },
+});
+
+const RewardColumn = styled.div({
+  display: 'grid',
+  textAlign: 'center',
+  '& > span:first-child': {
+    fontSize: '16px',
+    lineHeight: '20px',
+    marginBottom: '12px',
+  },
+  '& > span:last-child': {
+    fontFamily: 'Yapari-SemiBold',
+    fontSize: '16px',
+    lineHeight: '19.2px',
+    marginTop: '19px',
+    letterSpacing: '4%',
+  },
+});
+
+const Row = styled.div(({ columns, columnGap, justify }: { columns: string; columnGap: number; justify?: string }) => ({
   position: 'relative',
   display: 'grid',
   gridTemplateColumns: columns,
   columnGap,
   alignItems: 'center',
+  justifyContent: justify,
 }));
 
+const Rank = styled.span({
+  fontWeight: 600,
+  fontSize: '14px',
+  lineHeight: '14px',
+  padding: '6px 8px',
+  background: 'rgba(255, 255, 255, 0.2)',
+  borderRadius: '8px',
+  marginLeft: '4px',
+});
+
 export const TopWidget = ({
+  hasActiveTournament = false,
   amountOfDaysLeft = 1,
   onPlayAgain,
   tournamentTitle,
+  tournamentSubtitle,
   onTweet,
   score,
+  rank,
 }: TopWidgetProps): JSX.Element => {
   const { sdkUrl: cdnUrl, gamePortalUrl } = useConfigContext();
   const { address, isReady } = useWalletContext();
@@ -147,32 +224,84 @@ export const TopWidget = ({
   }, [address, gameInfo.name, onTweet, score]);
 
   return (
-    <WidgetWrapper layer={`${cdnUrl}/assets/layer.svg`} padding={[3, 3, 3, 3]}>
+    <WidgetWrapper layer={`${cdnUrl}/assets/layer.svg`} padding={[3, 3, 3, 3]} tournament={hasActiveTournament}>
       <RadialGradientBackGround />
       <Content>
-        <DaysLeft>{amountOfDaysLeft} day left</DaysLeft>
-        <NFTImage src={`${cdnUrl}/assets/nft.png`} />
-        <Text>
-          <Typography>{tournamentTitle}</Typography>
-          <UniqueNFT>TOP 20 WINS UNIQUE NFT</UniqueNFT>
-        </Text>
-        <Row columns={'146px 99px'} columnGap={8}>
-          <PlayAgain onClick={onPlayAgain}>
-            <div style={{ display: 'flex', alignItems: 'center' }}>
-              <RepeatIcon />
-              <PlayAgainText>Play Again</PlayAgainText>
-            </div>
-          </PlayAgain>
-          <TweetButton
-            disabled={!isReady || !address}
-            icon={<TwitterIcon color="#FFF" />}
-            variant="outlined"
-            onClick={handleShareClick}
-          >
-            Share
-          </TweetButton>
-        </Row>
-        <GamePortalButton onClick={handleOpenGamePortal}>Go to Cere game portal →</GamePortalButton>
+        {!hasActiveTournament ? (
+          <>
+            <DaysLeft>{amountOfDaysLeft} day left</DaysLeft>
+            <NFTImage src={`${cdnUrl}/assets/nft.png`} />
+            <Text>
+              <Typography>{tournamentTitle}</Typography>
+              <UniqueNFT>{tournamentSubtitle}</UniqueNFT>
+            </Text>
+            <Row columns={'146px 99px'} columnGap={8}>
+              <PlayAgain onClick={onPlayAgain}>
+                <div style={{ display: 'flex', alignItems: 'center' }}>
+                  <RepeatIcon />
+                  <PlayAgainText>Play Again</PlayAgainText>
+                </div>
+              </PlayAgain>
+              <TweetButton
+                disabled={!isReady || !address}
+                icon={<TwitterIcon color="#FFF" />}
+                variant="outlined"
+                onClick={handleShareClick}
+              >
+                Share
+              </TweetButton>
+            </Row>
+            <GamePortalButton onClick={handleOpenGamePortal}>Go to Cere game portal →</GamePortalButton>
+          </>
+        ) : (
+          <>
+            <UniqueNFT align="center" tournament>
+              {tournamentSubtitle}
+            </UniqueNFT>
+            <Typography align="center">{tournamentTitle}</Typography>
+            <DaysLeft tournament={hasActiveTournament}>{amountOfDaysLeft} day left</DaysLeft>
+            <RewardsRow>
+              <RewardColumn>
+                <span>1st prize</span>
+                <img src={`${cdnUrl}/assets/first-place-reward.svg`} alt="First place reward" />
+                <span>USDT</span>
+              </RewardColumn>
+              <RewardColumn>
+                <span>2nd prize</span>
+                <img src={`${cdnUrl}/assets/second-place-reward.svg`} alt="Second place reward" />
+                <span>USDT</span>
+              </RewardColumn>
+              <RewardColumn>
+                <span>3rd prize</span>
+                <img src={`${cdnUrl}/assets/third-place-reward.svg`} alt="Third place reward" />
+                <span>USDT</span>
+              </RewardColumn>
+            </RewardsRow>
+            <Typography align="center">
+              Your rank <Rank>{rank}</Rank>
+            </Typography>
+            <Row columns={'130px 130px'} columnGap={6} justify="center">
+              <PlayAgain onClick={onPlayAgain} tournament={hasActiveTournament}>
+                <div style={{ display: 'flex', alignItems: 'center' }}>
+                  <RepeatIcon />
+                  <PlayAgainText>Play Again</PlayAgainText>
+                </div>
+              </PlayAgain>
+              <TweetButton
+                tournament={hasActiveTournament}
+                disabled={!isReady || !address}
+                icon={<TwitterIcon color="#FFF" />}
+                variant="outlined"
+                onClick={handleShareClick}
+              >
+                Share
+              </TweetButton>
+            </Row>
+            <GamePortalButton tournament={hasActiveTournament} onClick={handleOpenGamePortal}>
+              Go to Cere game portal →
+            </GamePortalButton>
+          </>
+        )}
       </Content>
     </WidgetWrapper>
   );